+++ internal/client/notification_discord.go
@@ -17,7 +17,7 @@ func (c *Client) CreateDiscordChannel(spaceID string, commonParams NotificationC
 		Name:                     commonParams.Name,
 		IntegrationID:            commonParams.Integration.ID,
 		Rooms:                    commonParams.Rooms,
-		Alarms:                   commonParams.Alarms,
+		NotificationOptions:      commonParams.NotificationOptions,
 		RepeatNotificationMinute: commonParams.RepeatNotificationMinute,
 	}
 
@@ -73,7 +73,7 @@ func (c *Client) UpdateDiscordChannelByID(spaceID string, commonParams Notificat
 	reqBody := notificationRequestPayload{
 		Name:                     commonParams.Name,
 		Rooms:                    commonParams.Rooms,
-		Alarms:                   commonParams.Alarms,
+		NotificationOptions:      commonParams.NotificationOptions,
 		RepeatNotificationMinute: commonParams.RepeatNotificationMinute,
 	}