+++ internal/provider/notification_discord_channel_resource.go
@@ -36,7 +36,7 @@ type discordChannelResourceModel struct {
 	Enabled                  types.Bool   `tfsdk:"enabled"`
 	SpaceID                  types.String `tfsdk:"space_id"`
 	RoomsID                  types.List   `tfsdk:"rooms_id"`
-	Alarms                   types.String `tfsdk:"alarms"`
+	NotificationOptions      types.List   `tfsdk:"notifications"`
 	RepeatNotificationMinute types.Int64  `tfsdk:"repeat_notification_min"`
 	WebhookURL               types.String `tfsdk:"webhook_url"`
 	ChannelType              types.String `tfsdk:"channel_type"`
@@ -118,11 +118,14 @@ func (s *discordChannelResource) Create(ctx context.Context, req resource.Create
 	var roomsID []string
 	plan.RoomsID.ElementsAs(ctx, &roomsID, false)
 
+	var notificationOptions []string
+	plan.NotificationOptions.ElementsAs(ctx, &notificationOptions, false)
+
 	commonParams := client.NotificationChannel{
 		Name:                     plan.Name.ValueString(),
 		Integration:              *notificationIntegration,
 		Rooms:                    roomsID,
-		Alarms:                   plan.Alarms.ValueString(),
+		NotificationOptions:      notificationOptions,
 		Enabled:                  plan.Enabled.ValueBool(),
 		RepeatNotificationMinute: plan.RepeatNotificationMinute.ValueInt64(),
 	}
@@ -152,7 +155,7 @@ func (s *discordChannelResource) Create(ctx context.Context, req resource.Create
 	plan.Name = types.StringValue(notificationChannel.Name)
 	plan.Enabled = types.BoolValue(notificationChannel.Enabled)
 	plan.RoomsID, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.Rooms)
-	plan.Alarms = types.StringValue(notificationChannel.Alarms)
+	plan.NotificationOptions, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.NotificationOptions)
 	plan.RepeatNotificationMinute = types.Int64Value(notificationChannel.RepeatNotificationMinute)
 
 	diags = resp.State.Set(ctx, plan)
@@ -196,7 +199,7 @@ func (s *discordChannelResource) Read(ctx context.Context, req resource.ReadRequ
 	state.Name = types.StringValue(notificationChannel.Name)
 	state.Enabled = types.BoolValue(notificationChannel.Enabled)
 	state.RoomsID, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.Rooms)
-	state.Alarms = types.StringValue(notificationChannel.Alarms)
+	state.NotificationOptions, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.NotificationOptions)
 	state.RepeatNotificationMinute = types.Int64Value(notificationChannel.RepeatNotificationMinute)
 	state.WebhookURL = types.StringValue(notificationSecrets.URL)
 	state.ChannelType = types.StringValue(notificationSecrets.ChannelParams.Selection)
@@ -231,11 +234,14 @@ func (s *discordChannelResource) Update(ctx context.Context, req resource.Update
 	var roomsID []string
 	plan.RoomsID.ElementsAs(ctx, &roomsID, false)
 
+	var notificationOptions []string
+	plan.NotificationOptions.ElementsAs(ctx, &notificationOptions, false)
+
 	commonParams := client.NotificationChannel{
 		ID:                       plan.ID.ValueString(),
 		Name:                     plan.Name.ValueString(),
 		Rooms:                    roomsID,
-		Alarms:                   plan.Alarms.ValueString(),
+		NotificationOptions:      notificationOptions,
 		Enabled:                  plan.Enabled.ValueBool(),
 		RepeatNotificationMinute: plan.RepeatNotificationMinute.ValueInt64(),
 	}
@@ -265,7 +271,7 @@ func (s *discordChannelResource) Update(ctx context.Context, req resource.Update
 	plan.Name = types.StringValue(notificationChannel.Name)
 	plan.Enabled = types.BoolValue(notificationChannel.Enabled)
 	plan.RoomsID, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.Rooms)
-	plan.Alarms = types.StringValue(notificationChannel.Alarms)
+	plan.NotificationOptions, _ = types.ListValueFrom(ctx, types.StringType, notificationChannel.NotificationOptions)
 	plan.RepeatNotificationMinute = types.Int64Value(notificationChannel.RepeatNotificationMinute)
 
 	diags = resp.State.Set(ctx, plan)