+++ internal/provider/notification_discord_channel_resource_test.go
@@ -24,7 +24,7 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					space_id                = "%s"
 					rooms_id                = [netdata_room.test.id]
 					webhook_url             = "https://discord.com/api/webhooks/0000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"
-					alarms                  = "ALARMS_SETTING_ALL"
+					notifications           = ["CRITICAL","WARNING","CLEAR"]
 					channel_type            = "text"
 					repeat_notification_min = 30
 				}
@@ -35,7 +35,9 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "enabled", "true"),
 					resource.TestCheckResourceAttrSet("netdata_notification_discord_channel.test", "space_id"),
 					resource.TestCheckResourceAttrSet("netdata_notification_discord_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.0", "CRITICAL"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.1", "WARNING"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.2", "CLEAR"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "repeat_notification_min", "30"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "webhook_url", "https://discord.com/api/webhooks/0000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "channel_type", "text"),
@@ -53,7 +55,7 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					space_id                = "%s"
 					rooms_id                = [netdata_room.test.id]
 					webhook_url             = "https://discord.com/api/webhooks/0000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"
-					alarms                  = "ALARMS_SETTING_ALL"
+					notifications           = ["CLEAR","CRITICAL"]
 					channel_type            = "text"
 					repeat_notification_min = 60
 				}
@@ -64,7 +66,8 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "enabled", "true"),
 					resource.TestCheckResourceAttrSet("netdata_notification_discord_channel.test", "space_id"),
 					resource.TestCheckResourceAttrSet("netdata_notification_discord_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.0", "CLEAR"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.1", "CRITICAL"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "repeat_notification_min", "60"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "webhook_url", "https://discord.com/api/webhooks/0000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "channel_type", "text"),
@@ -82,7 +85,7 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					space_id       = "%s"
 					rooms_id       = null
 					webhook_url    = "https://discord.com/api/webhooks/1000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"
-					alarms         = "ALARMS_SETTING_ALL"
+					notifications  = ["CRITICAL","WARNING","CLEAR"]
 					channel_type   = "forum"
 					channel_thread = "thread"
 				}
@@ -93,7 +96,9 @@ func TestAccDiscordNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "enabled", "false"),
 					resource.TestCheckResourceAttrSet("netdata_notification_discord_channel.test", "space_id"),
 					resource.TestCheckNoResourceAttr("netdata_notification_discord_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.0", "CRITICAL"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.1", "WARNING"),
+					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "notifications.2", "CLEAR"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "repeat_notification_min", "0"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "webhook_url", "https://discord.com/api/webhooks/1000000000000/XXXXXXXXXXXXXXXXXXXXXXXX"),
 					resource.TestCheckResourceAttr("netdata_notification_discord_channel.test", "channel_type", "forum"),