+++ internal/provider/notification_pagerduty_channel_resource_test.go
@@ -23,7 +23,7 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					enabled                 = true
 					space_id                = "%s"
 					rooms_id                = [netdata_room.test.id]
-					alarms                  = "ALARMS_SETTING_ALL"
+					notifications           = ["CRITICAL","WARNING","CLEAR"]
 					alert_events_url        = "https://events.pagerduty.com/v2/enqueue"
 					integration_key         = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
 					repeat_notification_min = 30
@@ -35,7 +35,9 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "enabled", "true"),
 					resource.TestCheckResourceAttrSet("netdata_notification_pagerduty_channel.test", "space_id"),
 					resource.TestCheckResourceAttrSet("netdata_notification_pagerduty_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.0", "CRITICAL"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.1", "WARNING"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.2", "CLEAR"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "repeat_notification_min", "30"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alert_events_url", "https://events.pagerduty.com/v2/enqueue"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "integration_key", "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"),
@@ -52,7 +54,7 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					enabled                 = true
 					space_id                = "%s"
 					rooms_id                = [netdata_room.test.id]
-					alarms                  = "ALARMS_SETTING_ALL"
+					notifications           = ["CLEAR","CRITICAL"]
 					alert_events_url        = "https://events.pagerduty.com/v2/enqueue"
 					integration_key         = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
 					repeat_notification_min = 60
@@ -64,7 +66,8 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "enabled", "true"),
 					resource.TestCheckResourceAttrSet("netdata_notification_pagerduty_channel.test", "space_id"),
 					resource.TestCheckResourceAttrSet("netdata_notification_pagerduty_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.0", "CLEAR"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.1", "CRITICAL"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "repeat_notification_min", "60"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alert_events_url", "https://events.pagerduty.com/v2/enqueue"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "integration_key", "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"),
@@ -81,7 +84,7 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					enabled          = false
 					space_id         = "%s"
 					rooms_id         = null
-					alarms           = "ALARMS_SETTING_ALL"
+					notifications    = ["CRITICAL","WARNING","CLEAR"]
 					alert_events_url = "https://events.pagerduty.com/v2/enqueue"
 					integration_key  = "YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY"
 				}
@@ -92,7 +95,9 @@ func TestAccPagerdutyNotificationResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "enabled", "false"),
 					resource.TestCheckResourceAttrSet("netdata_notification_pagerduty_channel.test", "space_id"),
 					resource.TestCheckNoResourceAttr("netdata_notification_pagerduty_channel.test", "rooms_id.0"),
-					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alarms", "ALARMS_SETTING_ALL"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.0", "CRITICAL"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.1", "WARNING"),
+					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "notifications.2", "CLEAR"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "repeat_notification_min", "0"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "alert_events_url", "https://events.pagerduty.com/v2/enqueue"),
 					resource.TestCheckResourceAttr("netdata_notification_pagerduty_channel.test", "integration_key", "YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY"),